fix: UI/UX issues when no file associated to content
Previously, when no file was associated to the content field, the view would create a new record in attachment wire on page load. This resulted in Save/Cancel buttons being displayed immediately (due to wire changes) but the item wasn't in EDIT mode. Pressing cancel at this point would remove the new record from attachment wire but would result in the inability to add/update content if you press edit again (because there is no record in attachment wire).

This PR changes the behavior to create the new record in attachments when entering EDIT mode if there isn't already a record in attachment. This ensures that the user can always "edit" the content even if there isn't an existing file - keeping in mind that the content file can be deleted from anywhere in the system (e.g., recordmanager, another view, API call, etc.) so this view must account for the content file changing elsewhere.